+++ Changelog.md
@@ -3,6 +3,7 @@ OpenCore Changelog
 #### v0.7.4
 - Fixed Linux kernel sort order
 - Added Linux boot optional log detail
+- Fixed CPU core count detection for more legacy CPUs
 
 #### v0.7.3
 - Improved SSDT-PNLF compatibility with CFL+ graphics